+++ tools/castor/debug/Traceroute/cli.hh
@@ -39,9 +39,9 @@ public:
 	
 	bool get_ext();
 	int get_ttl();
-	int get_deadline();
+	float get_deadline();
 	bool contains_deadline();
-	int get_timeout();
+	float get_timeout();
 	AddressType get_address_type();
 	RouteType get_route_type();
 	SortType get_sort_type();
@@ -61,13 +61,13 @@ private:
 	int ttl = MAX_TTL;
 
 	// After x sec the program stops or if -x this attribute is ignored.
-	int deadline = -1;
+	float deadline = -1;
 
 	// Is the deadline set to a value other than -1
 	bool deadline_flag = false;
 
 	// Time to wait for a response in sec.
-	int timeout = 1;
+	float timeout = 1;
 
 	// Defines which addresses are to be shown.
 	AddressType address_type = AT_IP;
